+++ package.json
@@ -35,7 +35,6 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
-    "react-player": "^2.12.0",
     "react-redux": "^8.0.5",
     "react-router": "^6.8.1",
     "react-use": "^17.4.0",
